# docs/swagger.yaml
basePath: /
definitions:
  api.Accept:
    properties:
      flag:
        type: boolean
    type: object
  api.Audit:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
    type: object
  api.Destination:
    properties:
      address:
        type: string
      invert:
        type: string
    type: object
  api.Drop:
    properties:
      flag:
        type: boolean
    type: object
  api.ForwardPort:
    properties:
      port:
        type: string
      protocol:
        type: string
      toaddr:
        type: string
      toport:
        type: string
    type: object
  api.IcmpBlock:
    properties:
      name:
        type: string
    type: object
  api.IcmpType:
    properties:
      name:
        type: string
    type: object
...
  api.Limit:
    properties:
      value:
        type: string
    type: object
  api.Log:
    properties:
      level:
        type: string
      limit:
        $ref: '#/definitions/api.Limit'
      prefix:
        type: string
    type: object
  api.Mark:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
      set:
        type: string
    type: object
  api.Port:
    properties:
      port:
        type: string
      protocol:
        type: string
    type: object
  api.Protocol:
    properties:
      value:
        type: string
    type: object
...
  api.Reject:
    properties:
      flag:
        type: boolean
    type: object
  api.Rule:
    properties:
      accept:
        $ref: '#/definitions/api.Accept'
      audit:
        $ref: '#/definitions/api.Audit'
      destination:
        $ref: '#/definitions/api.Destination'
      drop:
        $ref: '#/definitions/api.Drop'
      family:
        type: string
      forwardport:
        $ref: '#/definitions/api.ForwardPort'
      icmpblock:
        $ref: '#/definitions/api.IcmpBlock'
      icmptype:
        $ref: '#/definitions/api.IcmpType'
      limit:
        $ref: '#/definitions/api.Limit'
      log:
        $ref: '#/definitions/api.Log'
      mark:
        $ref: '#/definitions/api.Mark'
      port:
        $ref: '#/definitions/api.Port'
      protocol:
        $ref: '#/definitions/api.Protocol'
      reject:
        $ref: '#/definitions/api.Reject'
      service:
        items:
          type: string
        type: array
      source:
        $ref: '#/definitions/api.Source'
      value:
        $ref: '#/definitions/api.Value'
    type: object
...
  api.Source:
    properties:
      address:
        type: string
      invert:
        type: string
      ipset:
        type: string
      mac:
        type: string
    type: object
...
  api.Value:
    properties:
      value:
        type: string
    type: object
...
  query.BatchForwardQuery:
    properties:
      delay:
        type: integer
      forwards:
        items:
          $ref: '#/definitions/query.ForwardQuery'
        type: array
    type: object
...
  query.BatchZoneQuery:
    properties:
      action_object:
        items:
          $ref: '#/definitions/query.ZoneDst'
        type: array
      delay:
        type: integer
    required:
    - action_object
    type: object
  query.ForwardQuery:
    properties:
      forward:
        $ref: '#/definitions/api.ForwardPort'
      ip:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - forward
    - ip
    type: object
...
  query.Query:
    properties:
      forward:
        $ref: '#/definitions/api.ForwardPort'
      ip:
        type: string
      port:
        $ref: '#/definitions/api.Port'
      rich:
        $ref: '#/definitions/api.Rule'
      service:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    type: object
...
  query.ZoneDst:
    properties:
      host:
        type: string
      zone:
        type: string
    required:
    - host
    - zone
    type: object
...
host: localhost:2952
info:
  contact: {}
  description: Uranus, distrubed firewall gateway.
  title: Uranus API
  version: 0.0.9
paths:
...
  /fw/v1/masquerade:
    delete:
      consumes:
      - application/json
      description: Disable masqerade on firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Disable masqerade on firewalld runtime.
      tags:
      - firewalld masquerade
    get:
      consumes:
      - application/json
      description: Get nat rule list on firewalld runtime.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get nat rule list on firewalld runtime.
      tags:
      - firewalld masquerade
    put:
      consumes:
      - application/json
      description: Enable masqerade on firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Enable masqerade on firewalld runtime.
      tags:
      - firewalld masquerade
  /fw/v1/nat:
    delete:
      consumes:
      - application/json
      description: Remove a nat rule at firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a nat rule at firewalld runtime.
      tags:
      - firewalld NAT
    get:
      consumes:
      - application/json
      description: Get nat rules at firewalld runtime
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get nat rules at firewalld runtime.
      tags:
      - firewalld NAT
    put:
      consumes:
      - application/json
      description: Add a nat rule at firewall runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a nat rule at firewall runtime.
      tags:
      - firewalld NAT
...
  /fw/v2/masquerade:
    delete:
      consumes:
      - application/json
      description: Disable masqerade at firewall permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Disable masqerade at firewall permanent.
      tags:
      - firewalld masquerade
    get:
      consumes:
      - application/json
      description: Get status of masqurade at firewalld permanent.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get status of masqurade at firewalld permanent.
      tags:
      - firewalld masquerade
    put:
      consumes:
      - application/json
      description: Enable masqerade at firewall permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Enable masqerade at firewall permanent.
      tags:
      - firewalld masquerade
  /fw/v2/nat:
    delete:
      consumes:
      - application/json
      description: Remove a nat rule at firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a nat rule at firewalld permanent.
      tags:
      - firewalld NAT
    get:
      consumes:
      - application/json
      description: Get nat rules at firewalld permanent.
      parameters:
      - description: ip
        in: query
        name: ip
        required: true
        type: string
      - description: zone
        in: query
        name: zone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get nat rules at firewalld permanent.
      tags:
      - firewalld NAT
    put:
      consumes:
      - application/json
      description: Add a nat rule at firewall permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a nat rule at firewall permanent.
      tags:
      - firewalld NAT
...
  /fw/v3/masquerade:
    delete:
      consumes:
      - application/json
      description: Disable masqerade on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchZoneQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Disable masqerade on firewalld runtime with delay timer.
      tags:
      - firewalld masquerade
    put:
      consumes:
      - application/json
      description: Enable masqerade on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchZoneQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Enable masqerade on firewalld runtime with delay timer.
      tags:
      - firewalld masquerade
  /fw/v3/nat:
    put:
      consumes:
      - application/json
      description: Add a NAT rule on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a NAT rule on firewalld runtime with delay timer.
      tags:
      - firewalld NAT
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"